Replace references to the Toulouse edition with Lyon
It was quite a while since Toulouse was replaced by Agen, Agen-1,
and Paris and several months since the Lyon edition was released
to it's about time we update the references in Sepia.

Since Toulouse the ID event has been added and it was included in
the "Test Outcome Analysis" pipeline actor role in overview.svg.
That might not be a perfect match but it's probably the best we've
got at this point.