Add SHOULD recommendation for publisher routing key schema (#9)
If all publishers choose their own routing key schema no consumers
will be able to use the routing key to filtering which events end
up in their queue. The proposed schema is very similar to what REMReM
uses except that it doesn't define the event family and use the actual
event name in the type field, i.e. instead of REMReM's

    eiffel.artifact.created.sometag.somedomain

Sepia will prescribe the following:

    eiffel.somefamily.EiffelArtifactCreatedEvent.sometag.somedomain

The point of this change relative to REMReM is that it's a trivial
standard for publishers to adhere to with no need for a table mapping
EiffelArtifactCreatedEvent to (artifact, created).