Uninline SVG images
For some reason the SVGs in index.html and test-orchestration.html
were both stored as separate files and inlined in the HTML. This made
it unnecessarily hard to edit the files and keep the copies in sync.

We instead use <object> elements to include the existing SVGs. Because
the images no longer are integrated into the document DOMs we had to
adjust the hyperlinks somewhat to use a new relative URL and a new
target attribute so the the link opens in the parent document rather
than in the SVG itself.

The only functional or visual change is that the hyperlinks within
the SVGs are no longer underlined.